Refactored Orion2 Injection Method
This was a revamp done in OrionDLL a long while back that was never reflected here. Since a mutex is created in CGameApp::InitInstance before anything else, it is the ideal location to inject all client hacks. In MapleStory, this is where WinMain is called. Unlike our previous methods of injection, this is guaranteed success on the first initial call, and will not trigger any access violation exceptions.

As a result, we don't need any alternative hooks like GetCurrentDirectoryA because nothing will work before NXL checks are called. Therefore, I've removed GetCurrentDirectoryA hooking as well as the MapleStory2_Bypass thread (we no longer need a thread running to validate this).