Fixed 0.0.0.0 IP Bug
I've updated hostname resolving and route address patching, as well as checking for if the IP is null. When the IP is null, but the gateway port is valid, it will log a warning to console and default the IP itself to the defined CLIENT_IP. However, if both IP and Port are both null, it will log it as an error and fail to continue.

Some explanation as to why this happens: in some version update, Nexon refactored the CGameApp::ParseCmdArgs function. When this happened, the IP and Port arg handling got changed. When you see the IP as 0.0.0.0, this means that no IP address argument was detected in the command line. In fact, the reason you see port 127 is simply because it's trying to parse the port first and yet our argument is the IP; thus, port 127 got resolved from the first set of IP digits.